+++ R/bittermelon-package.R
@@ -8,6 +8,8 @@
 #'                           [format.bm_bitmap()] and [print.bm_bitmap()].}
 #'     \item{bittermelon.bg}{Set new default `bg` argument value for
 #'                           [format.bm_bitmap()] and [print.bm_bitmap()].}
+#'     \item{bittermelon.col}{Set new default `col` argument value for
+#'                           [as.raster.bm_bitmap()] and [plot.bm_bitmap()].}
 #'     \item{bittermelon.compress}{Set new default `compress` argument value for
 #'                           [format.bm_bitmap()] and [print.bm_bitmap()].}
 #'     \item{bittermelon.monobit_path}{Set new default `monobit_path` argument value